requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.2
